Refactor: Update import paths and fix types
