+++ src/hooks/useRandomStudios.tsx
@@ -1,23 +1,23 @@
-import { useState, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { fetchRandomStudios } from '../reducers/studioSlice';
+import { useState, useEffect } from "react";
+import { useDispatch } from "react-redux";
+import { fetchRandomStudios } from "../reducers/studioReducer";
 
 const useRandomStudios = () => {
   const dispatch = useDispatch();
   const [data, setData] = useState([]);
-  const [status, setStatus] = useState('idle');
+  const [status, setStatus] = useState("idle");
   const [error, setError] = useState(null);
 
   useEffect(() => {
     const fetchData = async () => {
-      setStatus('loading');
+      setStatus("loading");
       try {
         const result = await dispatch(fetchRandomStudios()).unwrap();
         setData(result);
-        setStatus('succeeded');
+        setStatus("succeeded");
       } catch (err) {
         setError(err.message);
-        setStatus('failed');
+        setStatus("failed");
       }
     };
